Refactor: Egg Club API Typescript
