+++ src/app/(pages)/(club)/club/regular-club-sub/[id]/_components/NotificationList.tsx
@@ -4,21 +4,21 @@ import { ReactNode, useState } from "react";
 import { format, parseISO } from "date-fns";
 import ClubCard from "./ClubCard";
 import { ko } from "date-fns/locale";
-import { EggDay } from "@/types/eggday.types";
+import { EggDayWithEggDayMember } from "@/types/eggday.types";
 import Text from "@/components/uiComponents/TextComponents/Text";
-import { User } from "@/types/user.types";
+import { MemberInfo } from "@/types/user.types";
 
 interface NotificationListProps {
-  notificationData: EggDay[];
-  crewMembers: User[];
+  notificationData: EggDayWithEggDayMember[];
+  crewMembers: MemberInfo[];
   children?: ReactNode;
 }
 const NotificationList = ({ notificationData, crewMembers, children }: NotificationListProps) => {
   const [selectedDate, setSelectedDate] = useState("all");
 
   // 날짜별로 그룹화하는 함수
-  const groupDate = (notifications: EggDay[]) => {
-    return notifications.reduce((groups: { [key: string]: EggDay[] }, notification) => {
+  const groupDate = (notifications: EggDayWithEggDayMember[]) => {
+    return notifications.reduce((groups: { [key: string]: EggDayWithEggDayMember[] }, notification) => {
       // 날짜는 "MM.dd" 형식으로 변환
       const date = format(parseISO(notification.egg_day_date_time), "MM.dd");
 
@@ -57,46 +57,6 @@ const NotificationList = ({ notificationData, crewMembers, children }: Notificat
         </div>
       </div>
       <div className="self-stretch justify-start items-center gap-2 inline-flex">
-        {/* <button
-          onClick={() => setSelectedDate("all")}
-          className={`px-6 py-2 rounded-full border-2 ${
-            selectedDate === "all" ? "border-black bg-black text-white" : "border-gray-300"
-          }`}
-        >
-          전체
-        </button> */}
-
-        {/* <div className="flex gap-2 overflow-hidden overflow-x-auto scrollbar-hide w-full">
-          <button
-            onClick={() => setSelectedDate("all")}
-            className={`w-[41px] h-[68px]  rounded-[25px] border justify-center items-center gap-2.5 flex ${
-              selectedDate === "all" ? " border-neutral-800 border-2" : "  border-gray-100"
-            }`}
-          >
-            <Text
-              variant="body_medium-14"
-              className={`${selectedDate === "all" ? "text-neutral-800" : " text-gray-400"}`}
-            >
-              전체
-            </Text>
-          </button>
-
-          {dates.map((date) => (
-            <button
-              key={date}
-              onClick={() => setSelectedDate(date)}
-              className={`w-[41px] h-[68px] px-2 py-3 bg-white rounded-[25px] border flex-col justify-center items-center gap-1 inline-flex ${
-                selectedDate === date ? " border-neutral-800 border-2" : "border-gray-100"
-              }`}
-            >
-              <Text variant="body-10" className={`${selectedDate === date ? "text-neutral-800" : " text-gray-400"}`}>
-                {date}
-                {format(parseISO(groupedNotifications[date][0].egg_day_date_time), "eee", { locale: ko })}
-              </Text>
-            </button>
-          ))}
-        </div> */}
-
         <div className="flex gap-2 overflow-x-auto scrollbar-hide scroll-smooth pb-2">
           <button
             onClick={() => setSelectedDate("all")}
@@ -142,42 +102,6 @@ const NotificationList = ({ notificationData, crewMembers, children }: Notificat
       </div>
       {children}
     </div>
-
-    // <div className="flex flex-col gap-4">
-    //   <h2 className="text-xl font-bold">모임 일정</h2>
-
-    //   {/* 날짜 필터 버튼들 */}
-    //   <div className="flex gap-2 overflow-x-auto pb-2">
-    //     <button
-    //       onClick={() => setSelectedDate("all")}
-    //       className={`px-6 py-2 rounded-full border-2 ${
-    //         selectedDate === "all" ? "border-black bg-black text-white" : "border-gray-300"
-    //       }`}
-    //     >
-    //       전체
-    //     </button>
-
-    //     {dates.map((date) => (
-    //       <button
-    //         key={date}
-    //         onClick={() => setSelectedDate(date)}
-    //         className={`px-4 py-2 rounded-full border ${
-    //           selectedDate === date ? "border-black bg-black text-white" : "border-gray-300"
-    //         }`}
-    //       >
-    //         {date}
-    //         {format(parseISO(groupedNotifications[date][0].egg_day_date_time), "eee", { locale: ko })}
-    //       </button>
-    //     ))}
-    //   </div>
-
-    //   {/* 모임 카드 목록 */}
-    //   <div className="grid gap-4">
-    //     {filteredNotification().map((notification) => (
-    //       <ClubCard key={notification.egg_day_id} notification={notification} crewMembers={crewMembers} />
-    //     ))}
-    //   </div>
-    // </div>
   );
 };
