+++ src/app/(pages)/(club)/club/regular-club-sub/[id]/page.tsx
@@ -3,22 +3,19 @@ import TabLayout from "./_components/TabLayout";
 import ClubHeader from "./_components/ClubHeader";
 import HomeContent from "./_components/HomeContent";
 import RegularNotification from "./_components/RegularNotification";
-import { SubCategory } from "@/types/category.types";
-import { getEggClub, Member } from "@/types/안끝난거/eggclub.types";
-import { EggDay } from "@/types/eggday.types";
 import { MemberInfo } from "@/types/user.types";
 
-const OneTimeClubSubpage = async ({ params }: { params: { id: string } }) => {
+const OneTimeClubSubPage = async ({ params }: { params: { id: string } }) => {
   const regularClubId = Number(params.id);
 
-  const [memberData, notificationData, subCategories] = (await Promise.all([
+  const [memberData, notificationData, subCategories] = await Promise.all([
     getRegularMember(regularClubId),
     getRegularClubNotification(regularClubId),
     fetchSubCategories()
-  ])) as [Member[], EggDay[], SubCategory[]];
+  ]);
 
   // 클럽 정보만 추출
-  const clubInfo: getEggClub = memberData[0]?.egg_club;
+  const clubInfo = memberData[0]?.egg_club;
   // console.log("클럽인포:", clubInfo);
 
   // 일치하는 카테고리 찾기
@@ -62,4 +59,4 @@ const OneTimeClubSubpage = async ({ params }: { params: { id: string } }) => {
   );
 };
 
-export default OneTimeClubSubpage;
+export default OneTimeClubSubPage;
