+++ src/app/(pages)/(club)/club/regular-time/RegularContent.tsx
@@ -6,7 +6,7 @@ import { REGULAR_CLUB_CREATE } from "../_utils/localStorage";
 import Category from "../_components/regularClub/Category";
 import MemberType from "../_components/regularClub/MemberType";
 import { putRegularMember, putRepresentative } from "../_api/supabase";
-import { EggClubFormWithImageFile } from "@/types/안끝난거/eggclub.types";
+import { EggClubFormWithImageFile } from "@/types/eggclub.types";
 import ProgressBar from "../_components/ProgressBar";
 import { IoIosArrowBack } from "react-icons/io";
 import { Button } from "@/components/uiComponents/Button/ButtonCom";
